+++ src/modules/sbstudio/plugin/tasks/__init__.py
@@ -1,6 +1,13 @@
 from .base import Task
 from .initialization import InitializationTask
 from .light_effects import UpdateLightEffectsTask
+from .pixel_cache import InvalidatePixelCacheTask
 from .safety_check import SafetyCheckTask
 
-__all__ = ("Task", "InitializationTask", "SafetyCheckTask", "UpdateLightEffectsTask")
+__all__ = (
+    "Task",
+    "InitializationTask",
+    "InvalidatePixelCacheTask",
+    "SafetyCheckTask",
+    "UpdateLightEffectsTask",
+)